backend/requirements.txt
Django==4.2.1
django-cors-headers==4.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
daphne==4.0.0


